refactor(api): move update campaign to bounded context
