+++ api/src/prescription/campaign/infrastructure/repositories/campaign-administration-repository.js
@@ -3,6 +3,27 @@ import { knex } from '../../../../../db/knex-database-connection.js';
 import * as skillRepository from '../../../../../lib/infrastructure/repositories/skill-repository.js';
 import { Campaign } from '../../domain/read-models/Campaign.js';
 
+const get = async function (id) {
+  const campaign = await knex('campaigns').where({ id }).first();
+
+  if (!campaign) return null;
+
+  return new Campaign({
+    ...campaign,
+    organization: { id: campaign.organizationId },
+    targetProfile: { id: campaign.targetProfileId },
+    creator: { id: campaign.creatorId },
+  });
+};
+
+const update = async function (campaign) {
+  const editedAttributes = _.pick(campaign, ['name', 'title', 'customLandingPageText', 'ownerId']);
+
+  const [editedCampaign] = await knex('campaigns').update(editedAttributes).where({ id: campaign.id }).returning('*');
+
+  return new Campaign(editedCampaign);
+};
+
 const save = async function (campaigns, dependencies = { skillRepository }) {
   const trx = await knex.transaction();
   const campaignsToCreate = _.isArray(campaigns) ? campaigns : [campaigns];
@@ -56,4 +77,4 @@ const isCodeAvailable = async function (code) {
   return !(await knex('campaigns').first('id').where({ code }));
 };
 
-export { save, isCodeAvailable };
+export { save, update, get, isCodeAvailable };
