+++ api/tests/prescription/campaign/unit/application/campaign-administration-route_test.js
@@ -17,4 +17,18 @@ describe('Unit | Application | Router | campaign-administration-router ', functi
       expect(response.statusCode).to.equal(201);
     });
   });
+
+  describe('PATCH /api/campaigns/{id}', function () {
+    it('should return 400 with an invalid campaign id', async function () {
+      // given
+      const httpTestServer = new HttpTestServer();
+      await httpTestServer.register(moduleUnderTest);
+
+      // when
+      const response = await httpTestServer.request('PATCH', '/api/campaigns/invalid');
+
+      // then
+      expect(response.statusCode).to.equal(400);
+    });
+  });
 });
